separate sleep from web code since it is isomorphic
