+++ src/timer.ts
@@ -1,0 +1,3 @@
+export const sleep = (milliseconds: number): Promise<void> => {
+  return new Promise(resolve => setTimeout(resolve, milliseconds));
+};
